Move advances outside folder collection in the store
