+++ src/core/main/use-initialice-advances.js
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useDispatch } from "react-redux";
 //own
-import { getAdvancesFilm } from "store/collections/advances/get-advances";
+import { getAdvancesFilm } from "store/advances/get-advances";
 import { CATEGORY_STREAMING } from "constants";
 
 export default function useInitialiceAdvances(category = CATEGORY_STREAMING) {
